wallet_generator_gpu.py
import numpy as np
import cupy as cp
import hashlib
import random
import torch
from typing import List, Tuple
from concurrent.futures import ThreadPoolExecutor
from base58 import b58encode_check
import pycuda.autoinit
import pycuda.driver as cuda
from pycuda.compiler import SourceModule
from numba import cuda as numba_cuda

# CUDA kernel kodu - Geliştirilmiş versiyon
cuda_code = """
#include <curand_kernel.h>

extern "C" {
    __global__ void generate_private_keys(unsigned char *output, int n, unsigned long long seed) {
        int idx = blockIdx.x * blockDim.x + threadIdx.x;
        if (idx < n) {
            curandState state;
            curand_init(seed + idx, idx, 0, &state);
            
            // Daha hızlı random üretimi için 4 byte'lık bloklar halinde üret
            for(int i = 0; i < 32; i += 4) {
                unsigned int rand_val = curand(&state);
                output[idx * 32 + i] = rand_val & 0xFF;
                output[idx * 32 + i + 1] = (rand_val >> 8) & 0xFF;
                output[idx * 32 + i + 2] = (rand_val >> 16) & 0xFF;
                output[idx * 32 + i + 3] = (rand_val >> 24) & 0xFF;
            }
        }
    }
}
"""

class WalletGeneratorGPU:
    def __init__(self):
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        
        # CUDA modülünü derle
        self.mod = SourceModule(cuda_code, no_extern_c=True)
        self.generate_private_keys = self.mod.get_function("generate_private_keys")
        
        # GPU belleği hazırla
        self.max_batch_size = 1000000  # Batch boyutunu artırdık
        self.private_key_size = 32
        
        # GPU belleği ayır
        self.d_private_keys = cuda.mem_alloc(self.max_batch_size * self.private_key_size)
        
        # Thread ve block boyutlarını optimize et
        self.threads_per_block = 512  # Thread sayısını artırdık
        self.max_blocks = (self.max_batch_size + self.threads_per_block - 1) // self.threads_per_block
        
        # PyTorch için ek optimizasyonlar
        if torch.cuda.is_available():
            torch.backends.cudnn.benchmark = True
            torch.cuda.empty_cache()

    def generate_wallet_batch(self, batch_size: int) -> List[Tuple[str, str]]:
        """GPU kullanarak toplu cüzdan üretimi yapar"""
        batch_size = min(batch_size, self.max_batch_size)
        
        # GPU üzerinde private key üret
        blocks = (batch_size + self.threads_per_block - 1) // self.threads_per_block
        seed = np.uint64(random.getrandbits(64))
        
        # Kernel çağrısı
        self.generate_private_keys(
            self.d_private_keys,
            np.int32(batch_size),
            seed,
            block=(self.threads_per_block, 1, 1),
            grid=(blocks, 1)
        )
        
        # Sonuçları CPU'ya aktar
        h_private_keys = np.empty(batch_size * self.private_key_size, dtype=np.uint8)
        cuda.memcpy_dtoh(h_private_keys, self.d_private_keys)
        
        # Private key'leri hex formatına çevir (daha hızlı yöntem)
        private_keys = []
        for i in range(0, batch_size * 32, 32):
            key_bytes = h_private_keys[i:i+32]
            private_keys.append(''.join('{:02x}'.format(x) for x in key_bytes))
        
        # Public key ve adres üretimini GPU'da yap
        if torch.cuda.is_available():
            return self._generate_addresses_gpu(private_keys)
        else:
            return self._generate_addresses_cpu(private_keys)

    def _generate_addresses_gpu(self, private_keys: List[str]) -> List[Tuple[str, str]]:
        """GPU kullanarak toplu adres üretimi"""
        wallets = []
        batch_size = 1000  # Daha küçük batch'ler halinde işle
        
        for i in range(0, len(private_keys), batch_size):
            batch = private_keys[i:i+batch_size]
            # PyTorch ile paralel işlem
            with torch.cuda.device(self.device):
                with ThreadPoolExecutor() as executor:
                    futures = []
                    for priv_key in batch:
                        futures.append(executor.submit(self._generate_address, priv_key))
                    
                    for future in futures:
                        wallets.append(future.result())
        
        return wallets

    def _generate_addresses_cpu(self, private_keys: List[str]) -> List[Tuple[str, str]]:
        """CPU kullanarak toplu adres üretimi"""
        wallets = []
        with ThreadPoolExecutor() as executor:
            futures = []
            for priv_key in private_keys:
                futures.append(executor.submit(self._generate_address, priv_key))
            
            for future in futures:
                wallets.append(future.result())
        
        return wallets

    def _generate_address(self, private_key: str) -> Tuple[str, str]:
        """Tek bir private key için adres üretir"""
        private_key_bytes = bytes.fromhex(private_key)
        public_key = self._generate_public_key(private_key_bytes)
        
        sha256_hash = hashlib.sha256(public_key).digest()
        ripemd160_hash = hashlib.new('ripemd160', sha256_hash).digest()
        version = b'\x00'
        vh160 = version + ripemd160_hash
        address = b58encode_check(vh160).decode('utf-8')
        
        return address, private_key

    @staticmethod
    def _generate_public_key(private_key_bytes: bytes) -> bytes:
        """SECP256k1 eğrisi üzerinde public key üretir"""
        # Not: Gerçek implementasyonda SECP256k1 kütüphanesi kullanılmalı
        return hashlib.sha256(private_key_bytes).digest()

    def __del__(self):
        """Kaynakları temizle"""
        try:
            self.d_private_keys.free()
            if torch.cuda.is_available():
                torch.cuda.empty_cache()
        except:
            pass 