requirements.txt
bitarray==2.8.3
bitcoin==1.1.42
coincurve==18.0.0
ecdsa==0.18.0
base58==2.1.1
tqdm==4.66.1
concurrent-log-handler==0.9.24
psutil==5.9.8
numpy==1.26.4
numba==0.60.0
cupy-cuda12x==13.0.0
pycuda==2024.1
torch==2.2.0